web_service/requirements.txt
fastapi
uvicorn
supabase-py
openai
pydantic
python-dotenv
